### 2020-03-01 #371 @kenarab
RSQL
Submitting Author: Alejandro Baranek (@kenarab)  Leonardo Belen (@leobelen)
Repository:  https://github.com/rOpenStats/RSQL
Version submitted:   0.1.2

Version accepted: TBD

---

```
Package: RSQL
Title: Database Agnostic Package to Generate and Process 'SQL' Queries in R
Version: 0.1.1
    person("Alejandro", "Baranek",,"abaranek@dc.uba.ar", c("cre", "aut")),
    person("Leonardo","Belen",,"leobelen@gmail.com", c("aut")))
Maintainer: Alejandro Baranek <abaranek@dc.uba.ar>
Description: Allows the user to generate and execute select, insert, update and delete 'SQL' queries the underlying database without having to explicitly write 'SQL' code.
Collate:
    'sql-lib.R'
    'zzz.R'
Imports: lgr,
    R6,
    DBI,
    RSQLite
Suggests: knitr,
    covr

```

The objective to easily connect R code with databases without actually writing SQL code

All kind of audiences

dbplyr but it is focused in data retrieving using tidy syntax. With RSQL, you can read, but also  manipulate database code.
