### 2020-03-05 #371 @annakrystalli
Hello @kenarab and many thanks for your submission.

We have been discussing the packages and, given the variety of packages that perform similar functionality to yours, we are finding it difficult to decide whether it would be considered best in class because of the very minimal documentation.

Although you mention it has additional features to `dbplyr` in your submission, it is not clear from the README. Before we make a final decision, we would appreciate improvements to the documentation, demonstrating more of package functionality and how the package is different and more feature-complete than what exists already. Indeed being explicit about this in the docs would be useful for users in choosing the package, not just us. With that respect, a `pkgdown` documentation website would make it easier for users to browse documentation online. This is not necessary for accepting the package but I strongly recommend it.

Finally, can I ask that you please not add the rOpenSci review badge until the package has officially been accepted for review? Many thanks.

Any questions, just let me know.
