### 2020-03-05 #371 @kenarab
Hi @annakrystalli

We didn't find a similar package, if you know a package which can automatically generate SQL code using R code is really wellcome. dbplyr is focused in analysing data from a DB so it is focused on dplyr grammar and generates SELECTS statements. RSQL generates SELECT but, INSERT, UPDATES and DELETES. The expresiveness is limited yet but can be expanded with some work.

We will extend examples and documentation for justifying the opportunity to include this project in a so-rich community like rOpenSci.

One feature we want to explore using RSQL is to make a data.frame/data.table/tibble implementation which impacts directly on the database and not works in memory.

> Finally, can I ask that you please not add the rOpenSci review badge until the package has officially been accepted for review? Many thanks.

Badge configuration already removed from readme

Best, Ale.
