### 2020-03-05 #371 @annakrystalli
Thanks for the response @kenarab ! I'm confident that with more detailed documentation, it will be clear that the package is within scope. Look forward to seeing more of what the package can do 🙂

And yes @leobelen should be getting notifications of this thread because you tagged him in thhe initial comment but happy to tag in every comment too if you prefer?
