### 2020-03-17 #371 @noamross
⚠️⚠️⚠️⚠️⚠️

***In the interest of reducing load on reviewers and editors as we manage the
COVID-19 crisis, rOpenSci is temporarily pausing new submissions for software peer review
for 30 days (and possibly longer). Please check back here again after 17 April
for updates.***

***In this period new submissions will not be handled, nor new reviewers assigned.
Reviews and responses to reviews will be handled on a 'best effort' basis, but
no follow-up reminders will be sent.***

***Other rOpenSci community activities continue. We express our continued great
appreciation for the work of our authors and reviewers. Stay healthy and take
care of one other.***

***The rOpenSci Editorial Board***

⚠️⚠️⚠️⚠️⚠️
