### 2020-03-17 #371 @kenarab
It is great for me.
I´m putting effort this days working on COVID-19 trying to give clues from
international data to Southamerican outbreak, which is starting. There is
no hurry with RSQL process.

I´m building curated series from JHU for making estimations.
https://github.com/rOpenStats/COVID19

Maybe it is useful for you.

Best, Ale.
