### 2020-05-04 d9d34fd
using expect_equivalent() instead of expect_equal()
