+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: RSQL
 Type: Package
 Title: Database Agnostic Package to Generate and Process 'SQL' Queries in R
-Version: 0.1.1
+Version: 0.1.3
 Language: en-US
 Authors@R: c(
     person("Alejandro", "Baranek",,"abaranek@dc.uba.ar", c("cre", "aut")),
@@ -11,7 +11,7 @@ Description: Allows the user to generate and execute select, insert, update and 
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.0.2
+RoxygenNote: 7.1.0
 Collate:
     'sql-lib.R'
     'zzz.R'
