+++ NEWS.md
@@ -1,0 +1,6 @@
+RSQL 0.1.3
+============
+
+### NEW FEATURES
+
+* Initial public release
