+++ man/rm_vector_quotes.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/sql-lib.R
-\name{remove_quotes}
-\alias{remove_quotes}
+\name{rm_vector_quotes}
+\alias{rm_vector_quotes}
 \title{Removes quotes from data.frame columns}
 \usage{
-remove_quotes(text)
+rm_vector_quotes(text.vector)
 }
 \arguments{
 \item{text}{The text column to remove quotes from.}
