+++ NAMESPACE
@@ -6,6 +6,8 @@ export(createRSQL)
 export(getMtcarsdbPath)
 export(getPackageDir)
 export(parse_where_clause)
+export(rm_quotes)
+export(rm_vector_quotes)
 import(DBI)
 import(R6)
 import(lgr)
