### 2020-05-10 692c6fc
Corrections to DESCRIPTION and .travis.yml
