+++ .travis.yml
@@ -8,6 +8,7 @@ before_install:
 r_packages:
   - covr
   - lintr
+  - pkgdown
 env:
   - ON_TRAVIS = 1
 after_success:
