+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: '2.2'
+pkgdown: 1.5.1
+pkgdown_sha: ~
+articles:
+  rsql: rsql.html
+last_built: 2020-05-11T01:13Z
+
