### 2020-06-30 d699a7e
Implemented a apostrphe stuffing method for all commands
