+++ R/sql-lib.R
@@ -6,7 +6,6 @@
 # provide a tool to do it.
 
 
-# TODO where clause  with range as mpg > 1
 
 #' The class that provides the SQL functionality.
 #'
@@ -428,13 +427,25 @@ is_quoted <- function(text, quotes_symbols = "'") {
         ret <- FALSE
         while (!ret & i <= length(quotes_symbols)) {
           quotes <- quotes_symbols[i]
-          ret <- substr(text, 1, 1) == quotes & substr(text, nchar(text), nchar(text)) ==
-            quotes
+          ret <- substr(text, 1, 1) == quotes &
+                  substr(text, nchar(text), nchar(text)) == quotes
           i <- i + 1
         }
       }
     }
     ret
+}
+
+#' Stuff quote symbol from text
+#'
+#' @param text The unquoted string to stuff quotes from.
+stuff_quote <- function(unquoted.text, quote = "'"){
+  if (is_quoted(unquoted.text)){
+    stop(paste("stuff_quote function cannot be called with quoted text", unquoted.text))
+  }
+  quoted <- FALSE
+  stuffed.text <- gsub(quote, paste(quote, quote, sep = ""), unquoted.text)
+  stuffed.text
 }
 
 #' Removes the quotes from the string
@@ -450,15 +461,18 @@ dequote <- function(text) {
 #' @param text The string
 #' @param quotes The quotes
 re_quote <- function(text, quotes = "'") {
-    quote <- FALSE
-    if (!is_quoted(text, quotes_symbols = "\""))
+    if (!is.na(text)){
+      quote <- FALSE
+      if (!is_quoted(text, quotes_symbols = "\""))
         quote <- TRUE
-    if (is_quoted(text, quotes_symbols = "'")) {
+      if (is_quoted(text, quotes_symbols = "'")) {
         text <- dequote(text)
         quote <- TRUE
-    }
-    if (quote)
+      }
+      if (quote){
         text <- paste(quotes, text, quotes, sep = "")
+      }
+    }
     text
 }
 
@@ -478,10 +492,50 @@ add_quotes <- function(text) {
 #' @param quotes Quote characters
 #' @export
 rm_quotes <- function(text, quotes = "'") {
-    if (quotes == substr(text, 1, 1) & quotes == substr(text, nchar(text), nchar(text))) {
+    if (!is.na(text)){
+      unquoted.text <- text
+      if (quotes == substr(text, 1, 1) & quotes == substr(text, nchar(text), nchar(text))) {
         text <- substr(text, 2, nchar(text) - 1)
+        unquoted.text <- text
+      }
+      text <- stuff_quote(unquoted.text)
     }
     text
+}
+
+
+#' stuff quote characters in quoted or not quoted df for DSL or DML operations
+#'
+#' @param text.df Data Frame with corresponding values and fields as colnames
+#' @export
+stuff_df_quoted <- function(text.df){
+  if (!is.null(text.df)){
+    if (!is.data.frame(text.df)){
+      stop(paste("text.df must be a data.frame but is", class(text.df)[1]))
+    }
+    cols.quoted <- apply(text.df, MARGIN = 2, FUN = function(x)vapply(x, FUN = is_quoted, FUN.VALUE = logical(1)))
+    if (nrow(text.df) > 1){
+      cols.quoted.min <- apply(cols.quoted, MARGIN = 2, FUN = min)
+      cols.quoted.max <- apply(cols.quoted, MARGIN = 2, FUN = max)
+    }
+    else{
+      cols.quoted.min <- cols.quoted
+      cols.quoted.max <- cols.quoted
+    }
+    if (min(cols.quoted.min == cols.quoted.max) == 0){
[…]
