+++ man/stuff_quote.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sql-lib.R
+\name{stuff_quote}
+\alias{stuff_quote}
+\title{Stuff quote symbol from text}
+\usage{
+stuff_quote(unquoted.text, quote = "'")
+}
+\arguments{
+\item{text}{The unquoted string to stuff quotes from.}
+}
+\description{
+Stuff quote symbol from text
+}
