+++ .travis.yml
@@ -9,7 +9,6 @@ before_cache: Rscript -e 'remotes::install_cran("pkgdown")'
 r_packages:
   - covr
   - lintr
-  - pkgdown
 env:
   - ON_TRAVIS = 1
 after_success:
@@ -17,8 +16,7 @@ after_success:
   - Rscript -e 'library(covr); codecov()'
   - Rscript -e 'lintr::lint_package()'
   - Rscript -e 'pkgdown::build_site()'
-# pkgdown deploy not working
-#deploy:
-#  provider: script
-#  script: Rscript -e 'pkgdown::deploy_site_github(ssh_id = Sys.getenv("TRAVIS_DEPLOY_KEY", ""))'
-#  skip_cleanup: true
+deploy:
+  provider: script
+  script: Rscript -e 'pkgdown::deploy_site_github(ssh_id = Sys.getenv("TRAVIS_DEPLOY_KEY", ""))'
+  skip_cleanup: true
