+++ NAMESPACE
@@ -8,6 +8,7 @@ export(getPackageDir)
 export(parse_where_clause)
 export(rm_quotes)
 export(rm_vector_quotes)
+export(stuff_df_quoted)
 import(DBI)
 import(R6)
 import(lgr)
