+++ DESCRIPTION
@@ -11,7 +11,7 @@ Description: Allows the user to generate and execute select, insert, update and 
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.1.0
+RoxygenNote: 7.1.1
 Collate:
     'sql-lib.R'
     'zzz.R'
