+++ man/replaceNAwithNULL.Rd
@@ -1,0 +1,11 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sql-lib.R
+\name{replaceNAwithNULL}
+\alias{replaceNAwithNULL}
+\title{Replace NA with NULL in sql statement}
+\usage{
+replaceNAwithNULL(sql.code)
+}
+\description{
+Replace NA with NULL in sql statement
+}
