### 2020-07-02 5de9602
[bugfix] with no-data.frame values
