+++ R/sql-lib.R
@@ -276,7 +276,7 @@ RSQL.class <- R6::R6Class("RSQL", public = list(
     #' @param values The values (Not used. Included for compatibility)
     #' @param field_id The field of the serial id
     retrieve = function(table, fields_uk = names(values_uk), values_uk,
-                        fields = names(values), values,
+                        fields = names(values), values = NULL,
                                field_id = "id"){
       self$checkEntitiesNames(table, entity.type = "table")
       self$checkEntitiesNames(c(fields_uk, field_id), entity.type = "field")
@@ -508,36 +508,39 @@ rm_quotes <- function(text, quotes = "'") {
 #' @param text.df Data Frame with corresponding values and fields as colnames
 #' @export
 stuff_df_quoted <- function(text.df){
-  #debug
-  text.df <<- text.df
   if (!is.null(text.df)){
     if (!is.data.frame(text.df)){
       stop(paste("text.df must be a data.frame but is", class(text.df)[1]))
     }
-    cols.quoted <- apply(text.df, MARGIN = 2, FUN = function(x)vapply(x, FUN = is_quoted, FUN.VALUE = logical(1)))
-    if (nrow(text.df) > 1){
-      cols.quoted.min <- apply(cols.quoted, MARGIN = 2, FUN = min)
-      cols.quoted.max <- apply(cols.quoted, MARGIN = 2, FUN = max)
-    }
     else{
-      cols.quoted.min <- cols.quoted
-      cols.quoted.max <- cols.quoted
+      process <- nrow(text.df) > 0
     }
-    if (min(cols.quoted.min == cols.quoted.max) == 0){
-      unsound.columns <- which(cols.quoted.min != cols.quoted.max)
-      stop("unsound quoting strategy in columns", paste(names(text.df)[unsound.columns], collapse = ","))
-    }
-    for (i in seq_len(nrow(text.df))){
-      text.df[i, ] <- sapply(text.df[i, ], FUN = rm_quotes)
-      # Cannot apply vapply if NA value is in scope
-      #text.df[i, ] <- vapply(text.df[i, ], FUN = rm_quotes, FUN.VALUE = character(1))
-
-    }
-    for (j in seq_len(ncol(text.df))){
-      if (cols.quoted.min[j]){
-        text.df[, j] <- sapply(text.df[, j], FUN = re_quote)
+    if (process){
+      cols.quoted <- apply(text.df, MARGIN = 2, FUN = function(x)vapply(x, FUN = is_quoted, FUN.VALUE = logical(1)))
+      if (nrow(text.df) > 1){
+        cols.quoted.min <- apply(cols.quoted, MARGIN = 2, FUN = min)
+        cols.quoted.max <- apply(cols.quoted, MARGIN = 2, FUN = max)
+      }
+      else{
+        cols.quoted.min <- cols.quoted
+        cols.quoted.max <- cols.quoted
+      }
+      if (min(cols.quoted.min == cols.quoted.max) == 0){
+        unsound.columns <- which(cols.quoted.min != cols.quoted.max)
+        stop("unsound quoting strategy in columns", paste(names(text.df)[unsound.columns], collapse = ","))
+      }
+      for (i in seq_len(nrow(text.df))){
+        text.df[i, ] <- sapply(text.df[i, ], FUN = rm_quotes)
         # Cannot apply vapply if NA value is in scope
-        #text.df[, j] <- vapply(text.df[, j], FUN = re_quote, FUN.VALUE = character(1))
+        #text.df[i, ] <- vapply(text.df[i, ], FUN = rm_quotes, FUN.VALUE = character(1))
+
+      }
+      for (j in seq_len(ncol(text.df))){
+        if (cols.quoted.min[j]){
[…]
