+++ man/RSQL.class.Rd
@@ -381,7 +381,7 @@ Performs an insert on the database. This is a composite function
   fields_uk = names(values_uk),
   values_uk,
   fields = names(values),
-  values,
+  values = NULL,
   field_id = "id"
 )}\if{html}{\out{</div>}}
 }
