+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
   rsql: rsql.html
-last_built: 2020-05-11T01:13Z
+last_built: 2020-07-02T12:45Z
 
