### 2020-07-02 b191005
[bugfix] in testcases
