### 2020-07-02 9d6e470
Disabled Travis deployment for pkgdown
