+++ .travis.yml
@@ -16,7 +16,8 @@ after_success:
   - Rscript -e 'library(covr); codecov()'
   - Rscript -e 'lintr::lint_package()'
   - Rscript -e 'pkgdown::build_site()'
-deploy:
-  provider: script
-  script: Rscript -e 'pkgdown::deploy_site_github(ssh_id = Sys.getenv("TRAVIS_DEPLOY_KEY", ""))'
-  skip_cleanup: true
+# pkgdown in Travis is not yet working ok
+#deploy:
+#  provider: script
+  #script: Rscript -e 'pkgdown::deploy_site_github(ssh_id = Sys.getenv("TRAVIS_DEPLOY_KEY", ""))'
+#  skip_cleanup: true
