### 2020-07-02 99e66ce
needs_quotes for automatically adding quotes to corresponding types
