### 2020-07-02 59936dd
[bugfix] Filling double ' when character fields
