+++ R/sql-lib.R
@@ -468,6 +468,12 @@ dequote <- function(text) {
 #' @param text The string
 #' @param quotes The quotes
 re_quote <- function(text, quotes = "'") {
+    text.original <- text
+
+    # debug
+    text.original  <<- text.original
+
+    text <- as.character(text)
     if (!is.na(text)){
       quote <- FALSE
       if (!is_quoted(text, quotes_symbols = "\""))
@@ -480,6 +486,11 @@ re_quote <- function(text, quotes = "'") {
         text <- paste(quotes, text, quotes, sep = "")
       }
     }
+    if (substr(text, 1, 2) == "''" || substr(text, nchar(text)-1, nchar(text)) == "''"){
+      stop(paste("Text double quoted:", text, "original", text.original))
+    }
+    #debug
+    text <<- text
     text
 }
 
@@ -540,6 +551,7 @@ stuff_df_quoted <- function(text.df){
         unsound.columns <- which(cols.need.quotes.min != cols.need.quotes.max)
         stop("unsound quoting strategy in columns", paste(names(text.df)[unsound.columns], collapse = ","))
       }
+
       for (i in seq_len(nrow(text.df))){
         text.df[i, ] <- sapply(text.df[i, ], FUN = rm_quotes)
         # Cannot apply vapply if NA value is in scope
@@ -615,6 +627,9 @@ sql_gen_select <- function(select_fields, table,
 
 
     where_values.df <- stuff_df_quoted(where_values.df)
+    #debug
+    where_values.df <<- where_values.df
+
     separator <- ""
     sql_select_fields <- ""
     for (f in select_fields) {
@@ -650,69 +665,72 @@ sql_gen_select <- function(select_fields, table,
 #' @param where_fields The fields used in the where section
 #' @param where_values The values used in the where section
 sql_gen_where <- function(where_fields = names(where_values), where_values) {
-    ret <- ""
-    process <- !is.null(where_fields) & !is.null(where_values)
-    if (process){
-      if (is.data.frame(where_values)){
-        process <- nrow(where_values) > 0
+  #debug
+  print(where_values)
+  ret <- ""
+  process <- !is.null(where_fields) & !is.null(where_values)
+  if (process){
+    if (is.data.frame(where_values)){
+      process <- nrow(where_values) > 0
+    }
+    else{
+      if (!is.null(where_fields)){
+        stop(paste(gettext("sql_lib.no_where_values_specified", domain = "R-rsql")))
       }
-      else{
-        if (!is.null(where_fields)){
-          stop(paste(gettext("sql_lib.no_where_values_specified", domain = "R-rsql")))
-        }
-        if (!is.null(where_values)) {
-          stop(paste(gettext("sql_lib.no_where_values_specified", domain = "R-rsql")))
-        }
+      if (!is.null(where_values)) {
+        stop(paste(gettext("sql_lib.no_where_values_specified", domain = "R-rsql")))
       }
     }
-    if (process) {
-        # Asserts with values
-        if (!is.vector(where_fields))
-            stop(paste(gettext("sql_lib_where_files_has_to_be_a_vector",
-                               domain = "R-rsql"), str(where_fields)))
-        if (!is.list(where_values))
-            if (is.vector(where_values))
-                where_values <- data.frame(matrix(where_values,
-                                                  byrow = TRUE,
-                                                  ncol = length(where_values)),
-                  stringsAsFactors = FALSE)
-        if (length(where_fields) != ncol(where_values))
-            stop(paste(gettext("sql_lib.where_fields_num_not_eq_where_values_num", domain = "R-sql"), length(where_fields), "!=",
-                length(where_values), paste(where_fields, collapse = ","), paste(where_values,
-                  collapse = ",")))
[…]
