### 2020-07-02 23621ca
Removed debug info
