+++ NEWS.md
@@ -1,3 +1,11 @@
+RSQL 0.1.4
+============
+
+### NEW FEATURES
+
+- Now rsql stuffs ' within fields values, as ' is the quote symbol used by postgresql
+- All values to methods should be passed as data.frames/tibble
+
 RSQL 0.1.3
 ============
 
