### 2020-07-05 79eedf0
Minor correction in tests for applying dplyr::filter
