### 2020-07-15 #371 @geanders
@kenarab : We have re-set to our normal mode for reviews, and I see that we had left this review with a remaining question about documentation, in particular to help us figure out if this package meets our criteria for novelty. I wanted to check back in with you to see if you had any updates.

As a note, there are guidelines available in Chapter 5 of our [online book](https://devguide.ropensci.org/policies.html#aims-and-scope) on the point we're trying to assess before deciding whether to send the package out for review, and ideally the text of the package documentation (vignettes and / or README ) will highlight novelty or significant improvements over existing packages to help us assess that:

> "However, as we want packages in the rOpenSci suite to be our top recommendations for the tasks they perform, we aim to avoid duplication of functionality of existing R packages in any repo without significant improvements. An R package that replicates the functionality of an existing R package may be considered for inclusion in the rOpenSci suite if it significantly improves on alternatives in any repository (RO, CRAN, BioC) by being:

> - More open in licensing or development practices
> - Broader in functionality (e.g., providing access to more data sets, providing a greater suite of functions), but not only by duplicating additional packages
> - Better in usability and performance
> - Actively maintained while alternatives are poorly or no longer actively maintained
> - These factors should be considered as a whole to determine if the package is a significant improvement. A new package would not meet this standard only by following our package guidelines while others do not, unless this leads to a significant difference in the areas above.

> We recommend that packages highlight differences from and improvements over overlapping packages in their README and/or vignettes."
