### 2020-08-18 96f5d15
[bugfix] when NA as first field value in replaceNAwithNULL
