### 2020-12-02 2fc7801
Better coding when updating one field. Invariant checkings
