+++ man/re_quote_alt.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sql-lib.R
+\name{re_quote_alt}
+\alias{re_quote_alt}
+\title{This functions remove original quotes and sets validated quotes for corresponding db.
+If it had no quotes, will only put corresponding quotes symbols}
+\usage{
+re_quote_alt(text, quotes = "'")
+}
+\arguments{
+\item{text}{The string}
+
+\item{quotes}{The quotes}
+}
+\description{
+This functions remove original quotes and sets validated quotes for corresponding db.
+If it had no quotes, will only put corresponding quotes symbols
+}
