+++ man/check_fields_values.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/sql-lib.R
+\name{check_fields_values}
+\alias{check_fields_values}
+\title{Check fields and values are sound}
+\usage{
+check_fields_values(fields, values, min.length = 0)
+}
+\arguments{
+\item{fields}{Fields names to check}
+
+\item{values}{values to check}
+
+\item{min.length}{for vectors}
+}
+\description{
+Check fields and values are sound
+}
